hw4/requirements.txt
flask

# GCP Cloud Datastore library
google-cloud-datastore

# Python WSGI HTTP server (connects web servers such as nginx to app)
gunicorn

